extract console ui from local runtime
