+++ tests/launchers/test_local_launcher.py
@@ -10,19 +10,17 @@ from typing import Any, cast
 import pytest
 from loguru import logger as _base_logger
 
-from refiner.pipeline.data.shard import FilePart, Shard
-from refiner.pipeline import RefinerPipeline, read_csv, read_jsonl
-from refiner.cli.run.local import (
+import sys as local_run_sys
+from refiner.cli.ui.console import (
     LocalStageConsole,
-    sys as local_run_sys,
-    normalize_log_mode,
-    should_emit_worker_line,
-)
-from refiner.cli.run.local import (
     LocalStageSnapshot,
+    normalize_log_mode,
     run_local_stage_ui,
+    should_emit_worker_line,
     stream_local_stage_logs,
 )
+from refiner.pipeline.data.shard import FilePart, Shard
+from refiner.pipeline import RefinerPipeline, read_csv, read_jsonl
 from refiner.launchers.local import LaunchStats, LocalLauncher
 from refiner.pipeline.planning import PlannedStage, StageComputeRequirements
 from refiner.pipeline.sources.readers.base import BaseReader
@@ -253,7 +251,7 @@ def test_launch_local_streams_worker_logs_to_launcher_stdout(
         logger.info("processing row {}", row["x"])
         return {"x": int(row["x"]) + 1}
 
-    monkeypatch.setattr("refiner.cli.run.local.stdout_is_interactive", lambda: False)
+    monkeypatch.setattr("refiner.cli.ui.console.stdout_is_interactive", lambda: False)
 
     pipeline = read_jsonl(str(path)).map(logged_map)
     pipeline.launch_local(
@@ -270,7 +268,7 @@ def test_launch_local_streams_worker_logs_to_launcher_stdout(
 def test_local_stage_console_colors_timestamp_level_and_message(
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    monkeypatch.setattr("refiner.cli.run.local.stdout_is_interactive", lambda: True)
+    monkeypatch.setattr("refiner.cli.ui.console.stdout_is_interactive", lambda: True)
     console = LocalStageConsole(
         job_id="job-1",
         job_name="local-log-stream-demo",
@@ -322,7 +320,7 @@ def test_local_stage_console_colors_timestamp_level_and_message(
 def test_local_stage_console_omits_rundir_row_when_absent(
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    monkeypatch.setattr("refiner.cli.run.local.stdout_is_interactive", lambda: True)
+    monkeypatch.setattr("refiner.cli.ui.console.stdout_is_interactive", lambda: True)
     console = LocalStageConsole(
         job_id="job-1",
         job_name="cloud-attach-demo",
@@ -344,7 +342,7 @@ def test_local_stage_console_omits_rundir_row_when_absent(
 def test_local_stage_console_apply_snapshot_updates_stage_metadata(
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    monkeypatch.setattr("refiner.cli.run.local.stdout_is_interactive", lambda: False)
+    monkeypatch.setattr("refiner.cli.ui.console.stdout_is_interactive", lambda: False)
     console = LocalStageConsole(
         job_id="job-1",
         job_name="demo",
@@ -382,9 +380,9 @@ def test_local_stage_console_apply_snapshot_updates_stage_metadata(
 def test_local_stage_console_formats_system_lines_without_worker_prefix(
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    monkeypatch.setattr("refiner.cli.run.local.stdout_is_interactive", lambda: True)
+    monkeypatch.setattr("refiner.cli.ui.console.stdout_is_interactive", lambda: True)
     monkeypatch.setattr(
-        "refiner.cli.run.local.LocalStageConsole._render",
+        "refiner.cli.ui.console.LocalStageConsole._render",
         lambda *args, **kwargs: None,
     )
     console = LocalStageConsole(
@@ -409,9 +407,9 @@ def test_local_stage_console_formats_system_lines_without_worker_prefix(
 def test_local_stage_console_bounds_interactive_log_buffer(
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    monkeypatch.setattr("refiner.cli.run.local.stdout_is_interactive", lambda: True)
+    monkeypatch.setattr("refiner.cli.ui.console.stdout_is_interactive", lambda: True)
     monkeypatch.setattr(
-        "refiner.cli.run.local.LocalStageConsole._render",
+        "refiner.cli.ui.console.LocalStageConsole._render",
         lambda *args, **kwargs: None,
     )
     console = LocalStageConsole(
@@ -438,9 +436,9 @@ def test_local_stage_console_close_moves_cursor_below_frame(
 ) -> None:
     writes: list[str] = []
 
-    monkeypatch.setattr("refiner.cli.run.local.stdout_is_interactive", lambda: True)
+    monkeypatch.setattr("refiner.cli.ui.console.stdout_is_interactive", lambda: True)
     monkeypatch.setattr(
-        "refiner.cli.run.local.LocalStageConsole._write",
+        "refiner.cli.ui.console.LocalStageConsole._write",
         staticmethod(lambda text: writes.append(text)),
     )
 
@@ -467,9 +465,9 @@ def test_local_stage_console_close_prints_last_system_message_after_alt_screen(
 ) -> None:
     writes: list[str] = []
 
-    monkeypatch.setattr("refiner.cli.run.local.stdout_is_interactive", lambda: True)
+    monkeypatch.setattr("refiner.cli.ui.console.stdout_is_interactive", lambda: True)
     monkeypatch.setattr(
-        "refiner.cli.run.local.LocalStageConsole._write",
+        "refiner.cli.ui.console.LocalStageConsole._write",
         staticmethod(lambda text: writes.append(text)),
     )
 
@@ -502,7 +500,7 @@ def test_local_stage_console_write_squelches_broken_stdout(
             raise io.UnsupportedOperation()
 
     broken = _BrokenStdout()
-    monkeypatch.setattr("refiner.cli.run.local.sys.stdout", broken)
+    monkeypatch.setattr("refiner.cli.ui.console.sys.stdout", broken)
 
     with pytest.raises(BrokenPipeError):
         LocalStageConsole._write("hello")
@@ -517,21 +515,21 @@ def test_run_local_stage_ui_marks_failed_on_interrupt(
     system_messages: list[str] = []
     closed: list[bool] = []
 
-    monkeypatch.setattr("refiner.cli.run.local.stdout_is_interactive", lambda: False)
+    monkeypatch.setattr("refiner.cli.ui.console.stdout_is_interactive", lambda: False)
     monkeypatch.setattr(
-        "refiner.cli.run.local.stream_local_stage_logs",
+        "refiner.cli.ui.console.stream_local_stage_logs",
         lambda **kwargs: (_ for _ in ()).throw(KeyboardInterrupt()),
     )
     monkeypatch.setattr(
-        "refiner.cli.run.local.LocalStageConsole.set_status",
+        "refiner.cli.ui.console.LocalStageConsole.set_status",
         lambda self, status: statuses.append(status),
     )
     monkeypatch.setattr(
-        "refiner.cli.run.local.LocalStageConsole.emit_system",
+        "refiner.cli.ui.console.LocalStageConsole.emit_system",
         lambda self, message: system_messages.append(message),
     )
     monkeypatch.setattr(
-        "refiner.cli.run.local.LocalStageConsole.close",
+        "refiner.cli.ui.console.LocalStageConsole.close",
         lambda self: closed.append(True),
     )
 
@@ -565,7 +563,7 @@ def test_run_local_stage_ui_marks_failed_on_interrupt(
 def test_stream_local_stage_logs_skips_tails_for_none_mode(
     monkeypatch: pytest.MonkeyPatch, tmp_path: Path
 ) -> None:
-    monkeypatch.setattr("refiner.cli.run.local.stdout_is_interactive", lambda: False)
+    monkeypatch.setattr("refiner.cli.ui.console.stdout_is_interactive", lambda: False)
 
     class _Console:
         def emit(self, *, worker_id: str, line: str) -> None:
@@ -576,7 +574,7 @@ def test_stream_local_stage_logs_skips_tails_for_none_mode(
             return None
 
     monkeypatch.setattr(
-        "refiner.cli.run.local.LocalStageLogTail",
+        "refiner.cli.ui.console.LocalStageLogTail",
         lambda *args, **kwargs: (_ for _ in ()).throw(
             AssertionError("tail should not be created")
         ),
@@ -600,7 +598,7 @@ def test_stream_local_stage_logs_skips_tails_for_none_mode(
 def test_stream_local_stage_logs_only_tails_selected_worker_for_one_mode(
     monkeypatch: pytest.MonkeyPatch, tmp_path: Path
 ) -> None:
-    monkeypatch.setattr("refiner.cli.run.local.stdout_is_interactive", lambda: False)
+    monkeypatch.setattr("refiner.cli.ui.console.stdout_is_interactive", lambda: False)
     tailed: list[Path] = []
 
     class _FakeTail:
@@ -624,7 +622,7 @@ def test_stream_local_stage_logs_only_tails_selected_worker_for_one_mode(
             del force
             return None
 
-    monkeypatch.setattr("refiner.cli.run.local.LocalStageLogTail", _FakeTail)
+    monkeypatch.setattr("refiner.cli.ui.console.LocalStageLogTail", _FakeTail)
     ticks = 0
 
     def _is_stage_running() -> bool:
@@ -652,14 +650,14 @@ def test_local_stage_console_flushes_before_batched_truncation(
 ) -> None:
     renders: list[bool] = []
 
-    monkeypatch.setattr("refiner.cli.run.local.stdout_is_interactive", lambda: True)
+    monkeypatch.setattr("refiner.cli.ui.console.stdout_is_interactive", lambda: True)
     original_render = LocalStageConsole._render
 
     def _recording_render(self, *, force: bool = False) -> None:
         renders.append(force)
 
     monkeypatch.setattr(
-        "refiner.cli.run.local.LocalStageConsole._render",
+        "refiner.cli.ui.console.LocalStageConsole._render",
         _recording_render,
     )
 
@@ -683,7 +681,7 @@ def test_local_stage_console_flushes_before_batched_truncation(
         )
     finally:
         monkeypatch.setattr(
-            "refiner.cli.run.local.LocalStageConsole._render",
+            "refiner.cli.ui.console.LocalStageConsole._render",
             original_render,
         )
         console.close()
@@ -694,7 +692,7 @@ def test_local_stage_console_flushes_before_batched_truncation(
 def test_stream_local_stage_logs_errors_mode_keeps_traceback_continuations(
     monkeypatch: pytest.MonkeyPatch, tmp_path: Path
 ) -> None:
-    monkeypatch.setattr("refiner.cli.run.local.stdout_is_interactive", lambda: False)
+    monkeypatch.setattr("refiner.cli.ui.console.stdout_is_interactive", lambda: False)
     emitted: list[str] = []
 
     class _FakeTail:
@@ -729,7 +727,7 @@ def test_stream_local_stage_logs_errors_mode_keeps_traceback_continuations(
             del force
             return None
 
-    monkeypatch.setattr("refiner.cli.run.local.LocalStageLogTail", _FakeTail)
+    monkeypatch.setattr("refiner.cli.ui.console.LocalStageLogTail", _FakeTail)
     ticks = 0
 
     def _is_stage_running() -> bool:
